Fix compile warnings
Prevent signed/unsigned comparison issues by static_casting the int 'length' to size_t .
Remove the unused special_handling_tags constant from filter/string.cpp to clean up dead code and silence related warnings.
https://github.com/bibledit/cloud/issues/1095

+++ filter/mail.cpp
@@ -220,7 +220,7 @@ std::string filter_mail_address_name (std::string name)
 std::string filter_mail_limit_line_length_rfc5322(std::string body, const int length)
 {
   // If the HTML body length is within the maximum line length: Ready.
-  if (body.length() <= length)
+  if (body.length() <= static_cast<size_t>(length))
     return body;
 
   // Maximum number of iterations allowed: This prevents an infinite loop.
@@ -244,7 +244,7 @@ std::string filter_mail_limit_line_length_rfc5322(std::string body, const int le
     // Check whether the next new line after the caret is within the maximum line length.
     // If so update the caret position and go to next iteration.
     if (const auto nl_pos = body.find('\n', caret);
-        (nl_pos - caret) <= length) {
+        (nl_pos - caret) <= static_cast<size_t>(length)) {
       caret = nl_pos;
       continue;
     }
@@ -277,7 +277,7 @@ std::string filter_mail_limit_line_length_rfc5322(std::string body, const int le
     // Okay, the ">" or the "<" were not found: If needed:
     // 1. Add a new line in the body at the range end.
     // 2. Update the caret position.
-    if ((body.length() - caret) > length) {
+    if ((body.length() - caret) > static_cast<size_t>(length)) {
       const size_t range_end = std::min(caret + length, body.length());
       body.insert(range_end, "\n");
       caret = range_end;
